NagiosTable, NagiosStore, sorted & pretty
Added NagiosStore, which pulls from the nemo3 api. All host objects are
alphabetized and displayed in a table.

Also, the Table is beautified with a cool gradient.

Lastly, UpdateWindow added in. Mostly just a placeholder for someday
holding status updates and notifications.

LASTLY lastly, still a lot of stuff commented out.